refactor: product categories and subcategories
